websocket card events
Added a method for handling websocket events from the deck editor
view. At the moment, cards can be added to the deck, and these
changes will be saved to the DB, then reflected for everyone.
Most of the upcoming changes hinge on me formatting my json
document, which I don't wanna do, but I'll get around to it.